+++ app/src/main/java/com/example/pam_project/db/entities/ListEntity.java
@@ -9,14 +9,14 @@ import androidx.room.PrimaryKey;
 public class ListEntity {
 
     @PrimaryKey(autoGenerate = true)
-    public int id;
+    public long id;
 
     @ColumnInfo(name = "name")
     public String name;
 
     public int categoryId;
 
-    public ListEntity(int id, String name, int categoryId) {
+    public ListEntity(long id, String name, int categoryId) {
         this.id = id;
         this.name = name;
         this.categoryId = categoryId;
